add stickyButton , react-icons , wave svg , change a speed of scroll down animation.
