+++ package.json
@@ -43,6 +43,7 @@
     "react-day-picker": "^9.7.0",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.56.4",
+    "react-icons": "^5.5.0",
     "react-intersection-observer": "^9.16.0",
     "react-modal": "^3.16.3",
     "react-router-dom": "^7.6.0",
