+++ src/components/Govt/sections/HeroSection.tsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from "react";
-import { Calendar, ArrowDown, Eye } from "lucide-react";
+import { Calendar, CircleChevronDown, Eye } from "lucide-react";
 import { motion, AnimatePresence } from "framer-motion";
 import { GooeyText } from "../../ui/GooeyText";
 import { ParticleButton } from "@/components/ui/particle-button";
@@ -11,7 +11,7 @@ const heroSlides = [
 		title: "Over ₹10 Crore in Skill Projects Delivered.\nOne Trusted Partner.Seamless Execution.\nGuaranteed Impact.",
 		description: "Pan-India deployment. Tier 2 & 3 coverage.\nLMS, dashboards, assessments, and 13+ language trainers — all under one roof."
 	},
-	
+
 	{
 		bg: "url('/Govt-Images/banner1.JPG')",
 		title: (
@@ -25,14 +25,14 @@ const heroSlides = [
 					]}
 					morphTime={1}
 					cooldownTime={3}
-					className="font-bold text-4xl md:text-5xl lg:text-6xl text-center text-white relative z-10 whitespace-wrap inline-block mx-auto w-full"
+					className="font-bold text-4xl md:text-5xl lg:text-6xl text-center text-black relative z-10 whitespace-wrap inline-block mx-auto w-full"
 				/>
 			</div>
 		),
 		description: "Rareminds invites State Skill Development Missions (SSDMs), education departments,universities, \n& public sector initiatives to explore high-impact collaborations tailored for local needs."
 	},
 	{
-		bg: "url('/Govt-Images/banner1.JPG')",
+		bg: "url('/Govt-Images/banner2.JPG')",
 		title: (
 			<div className="w-full flex items-center justify-center min-h-[20px]">
 				<GooeyText
@@ -44,14 +44,14 @@ const heroSlides = [
 					]}
 					morphTime={1}
 					cooldownTime={3}
-					className="font-bold text-4xl md:text-5xl lg:text-6xl text-center text-white relative z-10 whitespace-wrap inline-block mx-auto w-full"
+					className="font-bold text-4xl md:text-5xl lg:text-6xl text-center text-black relative z-10 whitespace-wrap inline-block mx-auto w-full"
 				/>
 			</div>
 		),
 		description: "Rareminds invites State Skill Development Missions (SSDMs), education departments,universities, \n& public sector initiatives to explore high-impact collaborations tailored for local needs."
 	},
 	{
-		bg: "url('/Govt-Images/banner1.JPG')",
+		bg: "url('/Govt-Images/banner3.JPG')",
 		title: (
 			<div className="w-full flex items-center justify-center min-h-[20px]">
 				<GooeyText
@@ -63,13 +63,13 @@ const heroSlides = [
 					]}
 					morphTime={1}
 					cooldownTime={3}
-					className="font-bold text-4xl md:text-5xl lg:text-6xl text-center text-white relative z-10 whitespace-wrap inline-block mx-auto w-full"
+					className="font-bold text-4xl md:text-5xl lg:text-6xl text-center text-black relative z-10 whitespace-wrap inline-block mx-auto w-full"
 				/>
 			</div>
 		),
 		description: "Rareminds invites State Skill Development Missions (SSDMs), education departments,universities, \n& public sector initiatives to explore high-impact collaborations tailored for local needs."
 	},
-	
+
 ];
 
 const variants = {
@@ -120,9 +120,9 @@ export const HeroSection = () => {
 	}, []);
 
 	const paginate = (newDirection: number) => {
-    setDirection(newDirection);
-    setCurrentSlide(currentSlide + newDirection);
-  };
+		setDirection(newDirection);
+		setCurrentSlide(currentSlide + newDirection);
+	};
 
 	// Function to handle circular array indexing
 	function wrap(min: number, max: number, v: number) {
@@ -136,7 +136,7 @@ export const HeroSection = () => {
 		const interval = autoPlay ? setInterval(() => {
 			paginate(1);
 		}, isMobile ? 8000 : 12000) : null; // Shorter interval on mobile
-		
+
 		return () => {
 			if (interval) clearInterval(interval);
 		};
@@ -147,15 +147,15 @@ export const HeroSection = () => {
 		if (el) {
 			// Smoother scroll on mobile
 			const isMobile = window.innerWidth < 768;
-			el.scrollIntoView({ 
+			el.scrollIntoView({
 				behavior: "smooth",
-				block: isMobile ? "start" : "center" 
+				block: isMobile ? "start" : "center"
 			});
 		}
 	};
 
 	return (
-		<section className="fixed inset-0 h-[85%] md:h-[93%] flex items-center overflow-hidden flex-col sm:flex-row">
+		<section className="fixed inset-0  flex items-center overflow-hidden flex-col sm:flex-row">
 			<AnimatePresence initial={false} custom={direction} mode="popLayout">
 				<motion.div
 					key={slideIndex}
@@ -187,11 +187,11 @@ export const HeroSection = () => {
 							animate={{ opacity: 1 }}
 							exit={{ opacity: 0 }}
 							transition={{ duration: 0.6 }}
-							className="absolute inset-0 bg-black/30 backdrop-blur-[1px]"
+							className="absolute inset-0 bg-white/20 backdrop-blur-[2px]"
 						/>
 
 						{/* Content */}
-						<div className="relative w-full flex flex-col items-center justify-center px-4 sm:px-8 md:px-16 lg:px-24 text-white max-w-7xl mx-auto z-10">
+						<div className="relative w-full flex flex-col items-center justify-center px-4 sm:px-8 md:px-16 lg:px-24 text-black max-w-7xl mx-auto z-10">
 							<motion.h1
 								initial={{ opacity: 0, y: 30 }}
 								animate={{ opacity: 1, y: 0 }}
@@ -212,7 +212,7 @@ export const HeroSection = () => {
 								animate={{ opacity: 1, y: 0 }}
 								exit={{ opacity: 0, y: -30 }}
 								transition={{ duration: 0.6, delay: 0.1, ease: "easeOut" }}
-								className="text-sm sm:text-base md:text-lg text-gray-100 mb-6 sm:mb-8 text-center whitespace-pre-line max-w-4xl"
+								className="text-sm sm:text-base md:text-lg text-gray-800 mb-6 sm:mb-8 text-center whitespace-pre-line max-w-4xl"
 							>
 								{heroSlides[slideIndex].description}
 							</motion.p>
@@ -224,12 +224,18 @@ export const HeroSection = () => {
 								transition={{ duration: 0.6, delay: 0.2, ease: "easeOut" }}
 								className="flex text-xs md:text-base flex-wrap gap-2 sm:gap-4 justify-center"
 							>
-								<ParticleButton 
+								<ParticleButton
 									className="flex items-center p-2 md:px-4 md:py-2 bg-red-500/80 backdrop-blur rounded-full text-white shadow-sm shadow-red-400 border-2 border-red-300 hover:bg-red-600 transition-colors"
 									successDuration={1200}
 									size={"lg"}
 									particleColor="#ffffff"
 									particleSize={7}
+									onClick={() => {
+										const contactElement = document.getElementById('government-outcomes');
+										if (contactElement) {
+											contactElement.scrollIntoView({ behavior: 'smooth' });
+										}
+									}}
 								>
 									<Eye className="h-5 sm:h-6 w-5 sm:w-6 mr-2" />
 									View Our Government Training Portfolio
@@ -254,7 +260,7 @@ export const HeroSection = () => {
 			<motion.div
 				initial={{ opacity: 0 }}
 				animate={{ opacity: 1 }}
-				className="fixed right-4 sm:right-8 top-1/2 -translate-y-1/2 md:flex flex-col items-center gap-2 sm:gap-4 z-50 bg-black/20 px-2 py-3 sm:py-4 rounded-full hidden"
+				className="fixed right-4 sm:right-8 top-1/2 -translate-y-1/2 md:flex flex-col items-center gap-2 sm:gap-4 z-50 bg-black/30 px-2 py-3 sm:py-4 rounded-full hidden"
 			>
 				{heroSlides.map((_, index) => (
 					<button
@@ -263,20 +269,19 @@ export const HeroSection = () => {
 							setDirection(index > slideIndex ? 1 : -1);
 							setCurrentSlide(index);
 						}}
-						className={`w-2 sm:w-3 transition-all duration-300 rounded-full ${
-							slideIndex === index
+						className={`w-2 sm:w-3 transition-all duration-300 rounded-full ${slideIndex === index
 								? "bg-white h-4 sm:h-6 shadow-lg shadow-white/30"
 								: "bg-white/50 hover:bg-white/70 h-2 sm:h-3"
-						}`}
+							}`}
 						aria-label={`Go to slide ${index + 1}`}
 					/>
-					
+
 				))}
 			</motion.div>
 
 			{/* Scroll Down Icon */}
 			<div className="absolute bottom-2 md:bottom-6 left-10 sm:left-20 md:left-32 flex flex-col items-center z-50 animate-pulse bg-white/90 p-1 sm:p-2 backdrop-blur-sm shadow-md shadow-red-600 ring-2 sm:ring-4 ring-white/50 rounded-full"
-			onClick={scrollToGovOutcome}
+				onClick={scrollToGovOutcome}
 			>
 				<img
 					src="https://itvhjkgfafikpqmuunlh.supabase.co/storage/v1/object/public/images/Corporate/Recruitment/Index/Hero/scroll.png"
@@ -284,13 +289,13 @@ export const HeroSection = () => {
 					alt="Scroll down"
 					className="animate-spin-slow"
 				/>
-				<ArrowDown
-					size={30}
-					className=" absolute top-1/2 -translate-y-1/2 left-1/2 -translate-x-1/2 cursor-pointer"
+				<CircleChevronDown
+					size={35}
+					className="absolute top-1/2 -translate-y-1/2 left-1/2 -translate-x-1/2 cursor-pointer"
 					onClick={scrollToGovOutcome}
 				/>
 			</div>
-			
+
 		</section>
 	);
 };
